requirements.txt
matplotlib
pandas
Pillow
streamlit
colorsys
numpy
stegano